Reworking the footer(s) to add a SU footer.
It turned out to be a whole lot easier to just add a second footer than
to try to shoehorn a second footer into CKAN's single footer block.  A
necessary consequence of this decision, though, is that we are no longer
overriding anything, we're just statically defining the footer.

RE:#129